requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
docopt==0.6.2
filelock==3.0.12
gitignore==0.0.8
gunicorn==20.1.0
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2==2.9.1
pytz==2021.1
shell==1.0.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
virtualenv==20.7.1
virtualenv-clone==0.5.6
whitenoise==5.3.0
